requirements.txt
flask~=3.0.3
transformers~=4.41.2
torch
tensorflow
sentencepiece
tqdm~=4.66.4
Flask-WTF
wtforms~=3.1.2
PyMuPDF
google-generativeai
markdown2